# pre-allocate the capacity of the worker list