fix #3 
* 链表实际测过， 性能相差无几， 但内存占用高了几倍
* 初始化`Pool`和`PoolFunc`是给定了个默认的worker capacity
* 优化了下其他代码