@panjf2000 2018-07-25
@barryz 改的不错，不过有一点需要商榷，那个预分配的大小，之前benchmark我记得1000w也只需要用到7w多，这样的话预分配5w对于大部分场景其实有点浪费，你有没有用修改后的代码benchmark一下，每次allocs/op应该是变大了吧