@barryz 2018-07-25
@panjf2000  因为从压测结果来看， 1000W的场景goroutine池的作用很明显， 尤其是在内存开销方面，所以在选择默认worker数量时，取了一个相对较大的值。 下面是benchmark 对比

master:

```
goos: linux
goarch: amd64
BenchmarkGoroutineWithFunc-4           1        15079030287 ns/op       723120296 B/op  10176159 allocs/op
PASS
ok      command-line-arguments  15.106s
```

PR:
```
goos: linux
goarch: amd64
BenchmarkGoroutineWithFunc-4           1        15443376244 ns/op       720962888 B/op  10167850 allocs/op
PASS
ok      command-line-arguments  15.479s
```