@panjf2000 2018-07-25
@barryz 这样看起来预分配似乎没有太大的内存优势，相反在其他数量的任务量场景下可能还会有点浪费，这样吧，要不你先把预分配内存这一块的暂时移除，然后我合一下优化代码的部分，至于预分配内存这一块，后续再继续讨论下看看有没有能兼顾的办法